requirements.txt
fastapi
uvicorn
openai
python-dotenv
numpy
langchain-openai
thefuzz
sqlalchemy
psycopg2-binary
pydantic